Optimize unit test
