+++ tests/lora/conftest.py
@@ -148,7 +148,8 @@ def sql_lora_huggingface_id():
 
 
 @pytest.fixture(scope="session")
-def sql_lora_files(sql_lora_huggingface_id):
+def sql_lora_files():
+
     return snapshot_download(repo_id=sql_lora_huggingface_id)
